pi/requirements.txt
getmac==0.9.2
pingparsing==1.4.0
netifaces==0.11.0
psycopg2-binary==2.9.3
speedtest-cli==2.1.3
uvicorn==0.20.0
fastapi==0.92.0
requests==2.25.1
numpy==1.24.2
iperf3==0.1.11
psutil==5.9.2